Enhance ExMAS algorithm with configurable parallelism and improved logging
- Introduced `algorithmProcessCount` in `ExMasConfigGroup` to control parallelism in core ExMAS operations (pair generation and ride extension).
- Updated `PairGenerator`, `RidePostProcessor`, and `RideExtender` to utilize the new configuration for parallel processing.
- Enhanced logging in `PairGenerator` and `RideExtender` to provide progress updates with estimated time of arrival (ETA) during processing.
- Modified `MatsimNetworkCache` to support thread-local routing for improved performance and deterministic behavior.
- Adjusted `RunKelheimDemandExtraction` to accept command-line arguments for deterministic mode and process counts.
- Ensured consistent handling of absolute detour limits in `DrtRequestFactory`.
- Improved CSV output in `ExMasCsvWriter` by sorting rides and successors for better readability.
- Added offline input handling in `RunKelheimDemandExtraction` to avoid network issues during scenario setup.

+++ contribs/drt-demand-extraction/src/main/java/org/matsim/contrib/demand_extraction/demand/DrtRequestFactory.java
@@ -316,7 +316,7 @@ public class DrtRequestFactory {
 		
 		// Apply absolute detour cap if configured
 		if (exmasConfig.getMaxAbsoluteDetour() != null) {
-			maxAbsoluteDetour = Math.min(maxAbsoluteDetour, exmasConfig.getMaxAbsoluteDetour());
+			maxAbsoluteDetour = Math.min(maxAbsoluteDetour, (double) exmasConfig.getMaxAbsoluteDetour());
 		}
 		
 		double effectiveMaxDetourFactor = 1.0 + (maxAbsoluteDetour / drtAttrs.travelTime);
